+++ admin/src/client/custom.ts
@@ -1,0 +1,9 @@
+/**
+ * 自定义类型和工具函数，不会被代码生成器覆盖
+ */
+
+// 从types.ts导出所有类型
+export * from "./types"
+
+// 从utils.ts导出所有工具函数
+export * from "./utils"
