+++ admin/src/components/Common/ItemActionsMenu.tsx
@@ -2,7 +2,7 @@ import { IconButton } from "@chakra-ui/react"
 import { BsThreeDotsVertical } from "react-icons/bs"
 import { MenuContent, MenuRoot, MenuTrigger } from "../ui/menu"
 
-import type { ItemPublic } from "@/client"
+import type { ItemPublic } from "@/client/custom"
 import DeleteItem from "../Items/DeleteItem"
 import EditItem from "../Items/EditItem"
