+++ admin/src/components/Items/EditItem.tsx
@@ -11,7 +11,8 @@ import { useState } from "react"
 import { type SubmitHandler, useForm } from "react-hook-form"
 import { FaExchangeAlt } from "react-icons/fa"
 
-import { type ApiError, type ItemPublic, ItemsService } from "@/client"
+import { type ApiError, ItemsService } from "@/client"
+import { type ItemPublic } from "@/client/custom"
 import useCustomToast from "@/hooks/useCustomToast"
 import { handleError } from "@/utils"
 import {
