+++ admin/src/routes/_layout/items.tsx
@@ -11,7 +11,8 @@ import { createFileRoute, useNavigate } from "@tanstack/react-router"
 import { FiSearch } from "react-icons/fi"
 import { z } from "zod"
 
-import { type ItemPublic, ItemsService } from "@/client"
+import { ItemsService } from "@/client"
+import { type ItemPublic } from "@/client/custom"
 import { ItemActionsMenu } from "@/components/Common/ItemActionsMenu"
 import AddItem from "@/components/Items/AddItem"
 import PendingItems from "@/components/Pending/PendingItems"
